[tests] Added SSH support to test setup script

+++ test-messaging-bw-loop.py
@@ -1,12 +1,13 @@
+import logging
+import signal
+import subprocess
 import sys
 import time
-import subprocess
-import signal
-import logging
+
+
 import click
-from threading import Thread
 
-
+# from threading import Thread
 
 
 logging.basicConfig(
@@ -16,10 +17,16 @@ logging.basicConfig(
 logger = logging.getLogger(__name__)
 
 
+SSH_CONNECTION_TIMEOUT = 10
+
+
 class ProcessHasNotBeenCreated(Exception):
     pass
 
 class ProcessHasNotBeenStartedSuccessfully(Exception):
+    pass
+
+class ProcessHasNotBeenKilled(Exception):
     pass
 
 
@@ -38,9 +45,7 @@ def process_is_running(process):
         is_running = False
     return (is_running, returncode)
 
-
-
-def create_process(name, args):
+def create_process(name, args, via_ssh: bool=False):
     """ 
     name: name of the application being started
     args: process args
@@ -57,7 +62,7 @@ def create_process(name, args):
                 args, 
                 stdin =subprocess.PIPE,
                 stdout=subprocess.PIPE,
-                #stderr=subprocess.PIPE,
+                stderr=subprocess.PIPE,
                 universal_newlines=False,
                 creationflags=subprocess.CREATE_NEW_PROCESS_GROUP,
                 bufsize=1
@@ -66,8 +71,8 @@ def create_process(name, args):
             process = subprocess.Popen(
                 args, 
                 #stdin =subprocess.PIPE,
-                #stdout=subprocess.PIPE,
-                #stderr=subprocess.PIPE,
+                stdout=subprocess.PIPE,
+                stderr=subprocess.PIPE,
                 #universal_newlines=False,
                 bufsize=1
             )
@@ -76,7 +81,10 @@ def create_process(name, args):
 
     # Check that the process has started successfully and has not terminated
     # because of an error
-    time.sleep(1)
+    if via_ssh:
+        time.sleep(SSH_CONNECTION_TIMEOUT + 1)
+    else:
+        time.sleep(1)
     logger.debug('Checking that the process has started successfully: {}'.format(name))
     is_running, returncode = process_is_running(process)
     if not is_running:
@@ -87,9 +95,6 @@ def create_process(name, args):
     logger.debug('Started successfully')
     return process
 
-
-
-
 def cleanup_process(name, process):
     """ 
     Clean up actions for the process. 
@@ -97,29 +102,37 @@ def cleanup_process(name, process):
     Raises:
         ProcessHasNotBeenKilled
     """
+    # NOTE: There is a problem with terminating processes which use SSH 
+    # to run a command on a remote server. The problem is in SSH not 
+    # forwarding a signal (e.g., SIGINT, SIGTERM). As a result, SSH session 
+    # itself terminates and process.poll() returns None, however 
+    # an application started from a command continues to work on a remote server.
+    # The solution is to use -t option in order to allocate a pseudo-terminal. 
+    # See https://stackoverflow.com/questions/48419781/work-around-ssh-does-not-forward-signal
+    # for details. FIXME: Maybe it is reasonable to add additional check in
+    # clean-up actions that the process is not running on a remote server
+    # ps -A | grep [process_name]
+
+    # FIXME: However, there is a problem with wrong interpretation of carriage 
+    # (\r\n) from pseudo-terminal in this case. Check stdout, it is full of b'\r\n'.
 
     # FIXME: Signals may not work on Windows properly. Might be useful
     # https://stefan.sofa-rockers.org/2013/08/15/handling-sub-process-hierarchies-python-linux-os-x/
-    #process.stdout.close()
-    #process.stdin.close()
-    logger.debug('Terminating the process: {}'.format(name))
-    logger.debug('OS: {}'.format(sys.platform))
 
+    is_running, returncode = process_is_running(process)
+    if not is_running: 
+        logger.info(f'Process {name} is not running, no need to terminate')
+        return
+    
+    logger.info(f'Terminating the process: {name}')
+    # logger.info('OS: {}'.format(sys.platform))
     sig = signal.CTRL_C_EVENT if sys.platform == 'win32' else signal.SIGINT
-    #if sys.platform == 'win32':
-    #    if sig in [signal.SIGINT, signal.CTRL_C_EVENT]:
-    #        sig = signal.CTRL_C_EVENT
-    #    elif sig in [signal.SIGBREAK, signal.CTRL_BREAK_EVENT]:
-    #        sig = signal.CTRL_BREAK_EVENT
-    #    else:
-    #        sig = signal.SIGTERM
-
     process.send_signal(sig)
     for i in range(3):
         time.sleep(1)
         is_running, returncode = process_is_running(process)
         if not is_running: 
-            logger.debug('Terminated')
+            logger.info('Terminated')
             return
 
     # TODO: (For future) Experiment with this more. If stransmit will not 
@@ -130,14 +143,13 @@ def cleanup_process(name, process):
     # however process_is_running(process) becomes False
     is_running, _ = process_is_running(process)
     if is_running:
-        logger.debug('Killing the process: {}'.format(name))
+        logger.info(f'Killing the process: {name}')
         process.kill()
         time.sleep(1)
     is_running, _ = process_is_running(process)
     if is_running:
-        raise ProcessHasNotBeenKilled('{}, id: {}'.format(name, process.pid))
-    logger.debug('Killed')
-
+        raise ProcessHasNotBeenKilled(f'{name}, id: {process.pid}')
+    logger.info('Killed')
 
 def create_tshark(interface, port, output):
     args = ['tshark', '-i', interface, '-f', 'udp port {}'.format(port), '-s', '1500', '-w', output]
@@ -145,65 +157,151 @@ def create_tshark(interface, port, output):
 
 
 @click.command()
-@click.argument('dst_ip',   default="192.168.0.110")
-@click.argument('dst_port', default="4200")
+@click.argument('rcv_ssh_host')
+@click.argument('rcv_ssh_username')
+@click.argument('dst_host')
+@click.argument('dst_port')
+# help='Algorithm description'
 @click.argument('algdesc')
-@click.argument('pcapng')
+# help='File prefix - test scenario'
+@click.argument('scenario')
+# help='Network interface descriptor for tshark - eth0'
 @click.argument('iface')
 @click.option('--collect-stats', is_flag=True, help='Collect SRT statistics')
-def main(dst_ip, dst_port, algdesc, pcapng, iface, collect_stats):
-    common_args = ["./srt-test-messaging", "srt://{}:{}?sndbuf=12058624&smoother=live".format(dst_ip, dst_port), "",
-            "-msgsize", "1456", "-reply", "0", "-printmsg", "0"]
+def main(
+    rcv_ssh_host, 
+    rcv_ssh_username, 
+    dst_host, 
+    dst_port, 
+    algdesc, 
+    scenario, 
+    iface, 
+    collect_stats
+):
+            
+    rcv_path_to_srt = 'projects/srt/maxlovic'
+    snd_path_to_srt = '.'
 
-    if collect_stats:
-        common_args += ['-statsfreq', '1']
+    processes = []
+    try:
+        logger.info(f'Starting srt receiver on a remote machine: {rcv_ssh_host}')
+        # NOTE: It is important to add "-t" option in order for SSH 
+        # to transfer SIGINT, SIGTERM signals to the command
+        # NOTE: It is important to add "-o BatchMode=yes" option 
+        # in order to disable any kind of promt
+        # NOTE: It is important to add # "-o ConnectTimeout={SSH_CONNECTION_TIMEOUT}"
+        # option in case when the server is down not to wait and be able to check 
+        # quickly that the process has not been started successfully
+        rcv_srt_args = [
+            'ssh', 
+            '-t',
+            '-o', 'BatchMode=yes',
+            '-o', f'ConnectTimeout={SSH_CONNECTION_TIMEOUT}',
+            f'{rcv_ssh_username}@{rcv_ssh_host}',
+            f'{rcv_path_to_srt}/srt-test-messaging',
+            f'"srt://:{dst_port}?rcvbuf=12058624&smoother=live"',
+            '-msgsize', '1456',
+            '-reply', '0', 
+            '-printmsg', '0'
+        ]
+        rcv_srt_process = create_process(
+            'srt-test-messaging (rcv)', 
+            rcv_srt_args, 
+            True
+        )
+        processes.append(('srt-test-messaging (rcv)', rcv_srt_process))
+        logger.info('Started successfully')
 
-    pc_name = 'srt-test-messaging (SND)'
+        logger.info('Starting streaming with different values of bitrate')
 
-    for bitrate in range(50000000, 1100000000, 50000000):
-        # Calculate number of packets for 20 sec of streaming
-        # based on the target bitrate and packet size.
-        repeat = 20 * bitrate // (1456 * 8)
-        maxbw  = int(bitrate // 8 * 1.25)
-
-        pcapng_file = pcapng + "-alg-{}-blt-{}bps.pcapng".format(algdesc, bitrate)
-        tshark = create_tshark(interface = iface, port = dst_port, output = pcapng_file)
-        time.sleep(3)
-
-        args = common_args + ["-bitrate", str(bitrate), "-repeat", str(repeat)]
+        # TODO: Add quotes to srt uri
+        snd_common_srt_args = [
+            f'{snd_path_to_srt}/srt-test-messaging', 
+            f'srt://{dst_host}:{dst_port}?sndbuf=12058624&smoother=live',
+            "",
+            '-msgsize', '1456',
+            '-reply', '0', 
+            '-printmsg', '0'
+        ]
         if collect_stats:
-            stats_file = pcapng + "-alg-{}-blt-{}bps.csv".format(algdesc, bitrate)
-            args += ['-statsfile', stats_file]
-        args[1] += "&maxbw={}".format(maxbw)
-        logger.info("Starting with bitrate {}, repeat {}".format(bitrate, repeat))
-        snd_srt_process = create_process(pc_name, args)
+            snd_common_srt_args += ['-statsfreq', '1']
 
-        sleep_s = 20
-        is_running = True
-        i = 0
-        while is_running:
-            is_running, returncode = process_is_running(snd_srt_process)
-            if is_running:
-                time.sleep(sleep_s)
-                sleep_s = 1  # Next time sleep for 1 second to react on the process finished.
-                i += 1
 
-        logger.info("Done")
+        # for bitrate in range(50000000, 1100000000, 50000000):
+        for bitrate in range(5000000, 8000000, 1000000):
+            # Calculate number of packets for 20 sec of streaming
+            # based on the target bitrate and packet size.
+            repeat = 20 * bitrate // (1456 * 8)
+            maxbw  = int(bitrate // 8 * 1.25)
+
+            pcapng_file = scenario + "-alg-{}-blt-{}bps.pcapng".format(algdesc, bitrate)
+            tshark = create_tshark(
+                interface=iface, 
+                port=dst_port, 
+                output=pcapng_file
+            )
+            processes.append(('tshark', tshark))
+            time.sleep(3)
+
+            args = snd_common_srt_args + ["-bitrate", str(bitrate), "-repeat", str(repeat)]
+            if collect_stats:
+                stats_file = scenario + "-alg-{}-blt-{}bps.csv".format(algdesc, bitrate)
+                args += ['-statsfile', stats_file]
+            args[1] += "&maxbw={}".format(maxbw)
+            logger.info("Starting streaming with bitrate {}, repeat {}".format(bitrate, repeat))
+            snd_srt_process = create_process('srt-test-messaging (snd)', args)
+            processes.append(('srt-test-messaging (snd)', snd_srt_process))
+
+            sleep_s = 20
+            is_running = True
+            i = 0
+            while is_running:
+                is_running, returncode = process_is_running(snd_srt_process)
+                if is_running:
+                    time.sleep(sleep_s)
+                    sleep_s = 1  # Next time sleep for 1 second to react on the process finished.
+                    i += 1
+
+            logger.info("Done")
+            time.sleep(3)
+            cleanup_process("tshark", tshark)
+            if i >= 5:
+                logger.info("Waited {} seconds. {} is considered as max BW".format(20 + i, bitrate))
+                break
+
+        logger.info('Stopping srt receiver')
         time.sleep(3)
-        cleanup_process("tshark", tshark)
-        if i >= 5:
-            logger.info("Waited {} seconds. {} is considered as max BW".format(20 + i, bitrate))
-            break
+        cleanup_process('srt-test-messaging (rcv)', rcv_srt_process)
+    except KeyboardInterrupt:
+        logger.info('KeyboardInterrupt - Not implemented yet')
+    except (
+        ProcessHasNotBeenCreated,
+        ProcessHasNotBeenStartedSuccessfully,
+        # ProcessHasNotBeenTerminated,
+        ProcessHasNotBeenKilled,
+    ) as error:
+        logger.info(
+            f'Exception occured ({error.__class__.__name__}): {error}'
+        )
+    finally:
+        logger.info('Terminating running processes')
+        # TODO: Add go_further in a right way
+        # go_further = True    
+        if len(processes) == 0:
+            logger.info('Nothing to terminate')
+            return
 
-    # Start transmission in file mode
-    #args = ["srt-test-messaging", "srt://192.168.0.7:4200", "",
-    #        "-repeat", "500", "-printmsg", "0"]
-
+        for name, process in reversed(processes):
+            try:
+                cleanup_process(name, process)
+            except:
+                # TODO: Collect the information regarding non killed processes
+                # and perfom additional clean-up actions
+                # Exceptions: ProcessIsNotKilled and others
+                print('The next experiment can not be done further')
+                # go_further = False
+        # return go_further
 
 
 if __name__ == '__main__':
     main()
-
-
-
-
